+++ src/default.config.ts
@@ -5,6 +5,7 @@ export const DEFAULT_CONFIG: Config = {
   migrationsDir: 'migrations',
   defaultTemplate: 'es6',
   typescript: true,
+  outDir: './dist',
   logger: {
     folder: './migrations-log',
     up: {