+++ src/helpers/typescript-builder.ts
@@ -1,12 +1,14 @@
 import { spawn } from 'child_process';
 
-export const TranspileTypescript = (paths: string[]) => {
+export const TranspileTypescript = (paths: string[], outDir: string) => {
   return new Promise(resolve => {
     const child = spawn('npx', [
       'gapi',
       'build',
       '--glob',
-      `${paths.toString()}`
+      `${paths.toString()}`,
+      '--outDir',
+      outDir
     ]);
     // child.stdout.pipe(process.stdout);
     child.stderr.pipe(process.stderr);