+++ src/injection.tokens.ts
@@ -35,6 +35,7 @@ export interface Config {
       useNewUrlParser: true;
     };
   };
+  outDir: string;
   migrationsDir: string;
   changelogCollectionName: string;
   logger: LoggerConfig;