+++ xmigrate.js
@@ -1,26 +1,13 @@
 module.exports = async () => {
   return {
-    changelogCollectionName: 'migrations',
-    migrationsDir: 'migrations',
     defaultTemplate: 'typescript',
+    outDir: './dist',
     typescript: true,
     logger: {
-      folder: './migrations-log',
-      up: {
-        success: 'up.success.log',
-        error: 'up.error.log'
-      },
-      down: {
-        success: 'down.success.log',
-        error: 'down.error.log'
-      }
+      folder: './migrations-log'
     },
     mongodb: {
-      url: 'mongodb://localhost:27017',
-      databaseName: 'test',
-      options: {
-        useNewUrlParser: true
-      }
-    },
+      databaseName: 'test'
+    }
   };
 };