+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 * v0.7.13
 - Full typescript support! Files are Transpiled using ParcelJS insternally inside `@gapi/cli`
 - 500 Migrations run with Typescript > Javascript > Execution in 13 seconds!
+- Users can now choose `typescript` outFolder with config `outDir` defaults to `./dist`
+- Connections for `mongoose` and `mongodb` can be closed forcefully
+- Tests for specific scenarios added
+- Modified `createWithTemplate` method inside `MigrationService` to return value with `/` instead of empty line since it was breaking consistency
+- Introduced inside tests(For now) custom template providing string example: `xmigrate create mytemplate --typescript true --raw true --template 'oh my god i have a template'` later in version v0.7.14 will have this capabilities
 
 * v0.7.8
 - Changed `db` to `client` since we are appending `MongoClient` instead of database